#31 Realm에 데이터를 저장하는 프로세스를 ResultsViewController로 이동.
Realm에 데이터를 Write하는 프로세스를 메서드로 독립.
