+++ Yetda_iOS/Source/Views/StartView/StartViewController.swift
@@ -45,47 +45,6 @@ class StartViewController: BaseViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         
-//        let presents = Presents()
-//        presents.name = "향수"
-//        presents.tags.append("뷰티")
-//        presents.tags.append("실용")
-        
-        let results = Results()
-        
-        let gift1 = Gift()
-        gift1.name = "향수"
-        gift1.tags.append("뷰티")
-        gift1.tags.append("실용")
-        
-        let gift2 = Gift()
-        gift2.name = "꽃다발"
-        gift2.tags.append("여성")
-        gift2.tags.append("감성")
-        
-        let receiver = Receiver()
-        receiver.name = "May"
-        receiver.gifts.append(gift1)
-        receiver.gifts.append(gift2)
-        
-        results.receivers.append(receiver)
-        
-        do {
-            let realm = try Realm()
-            try realm.write {
-                realm.add(results)
-            }
-            let outcome = realm.objects(Results.self)
-            print(outcome[0].receivers)
-            
-        } catch let error as NSError {
-            // handle error
-            print(error)
-        }
-        
-//        try! realm.write{
-//            realm.add(presents)
-//        }
-        
         
         //        let settings = FirestoreSettings()
         //        Firestore.firestore().settings = settings
